# api-gateway/docs/swagger.yaml
definitions:
...
  order_proto.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/order_proto.OrderItem'
        type: array
      user_id:
        type: string
    type: object
  order_proto.CreateOrderResponse:
    properties:
      message:
        type: string
      order_id:
        type: string
      status:
        type: string
      total_amount:
        type: number
      unavailable_products:
        items:
          $ref: '#/definitions/order_proto.UnavailableProduct'
        type: array
    type: object
  order_proto.GetAllOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/order_proto.Order'
        type: array
    type: object
  order_proto.GetOrderResponse:
    properties:
      order:
        $ref: '#/definitions/order_proto.Order'
    type: object
  order_proto.GetOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/order_proto.Order'
        type: array
    type: object
  order_proto.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order_proto.OrderItem'
        type: array
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  order_proto.OrderItem:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  order_proto.UnavailableProduct:
    properties:
      available_quantity:
        type: integer
      product_id:
        type: string
      requested_quantity:
        type: integer
    type: object
  order_proto.UpdateOrderStatusRequest:
    properties:
      order_id:
        type: string
      status:
        type: string
    type: object
  order_proto.UpdateOrderStatusResponse:
    properties:
      order:
        $ref: '#/definitions/order_proto.Order'
    type: object
  payment_proto.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  payment_proto.PaymentRequest:
    properties:
      amount:
        type: number
      order_id:
        type: string
      user_id:
        type: string
    type: object
  payment_proto.PaymentResponse:
    properties:
      message:
        type: string
    type: object
  product_proto.CreateProductReq:
    properties:
      description:
        type: string
      discount_price:
        type: number
      name:
        type: string
      price:
        type: number
      stock_quantity:
        type: integer
    type: object
  product_proto.CreateProductRes:
    properties:
      message:
        type: string
      product_id:
        type: string
    type: object
  product_proto.ListProductsRes:
    properties:
      product:
        items:
          $ref: '#/definitions/product_proto.Product'
        type: array
    type: object
  product_proto.Product:
    properties:
      description:
        type: string
      discount_price:
        type: number
      name:
        type: string
      price:
        type: number
      product_id:
        type: string
      stock_quantity:
        type: integer
    type: object
  product_proto.UpdateProductRes:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Flash-Sale
  version: "1.0"
paths:
...
  /order/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order_proto.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete an order
      tags:
      - order
    get:
      consumes:
      - application/json
      description: Get details of a specific order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order_proto.GetOrderResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get order by ID
      tags:
      - order
  /order/create:
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Order request body
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order_proto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order_proto.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new order
      tags:
      - order
  /order/status:
    put:
      consumes:
      - application/json
      description: Update the status of an order
      parameters:
      - description: Update order status request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order_proto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order_proto.UpdateOrderStatusResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update order status
      tags:
      - order
  /order/user/{id}:
    get:
      consumes:
      - application/json
      description: Get all orders made by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order_proto.GetOrdersResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all orders of a user
      tags:
      - order
  /orders:
    get:
      consumes:
      - application/json
      description: Get a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order_proto.GetAllOrdersResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all orders
      tags:
      - order
  /payment/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific payment by its ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment_proto.Payment'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get payment by ID
      tags:
      - payment
  /payment/process:
    post:
      consumes:
      - application/json
      description: Process a new payment
      parameters:
      - description: Payment request body
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/payment_proto.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment_proto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Process a payment
      tags:
      - payment
  /product:
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: Product request body
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/product_proto.CreateProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product_proto.CreateProductRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new product
      tags:
      - product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product_proto.UpdateProductRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete a product by ID
      tags:
      - product
    get:
      consumes:
      - application/json
      description: Get details of a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product_proto.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get product by ID
      tags:
      - product
    put:
      consumes:
      - application/json
      description: Update information of a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product update request body
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/product_proto.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product_proto.UpdateProductRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update product details
      tags:
      - product
  /products:
    get:
      consumes:
      - application/json
      description: Get a list of all available products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product_proto.ListProductsRes'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all products
      tags:
      - product
...
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"